feat: deploy all contracts to Base & OP Sepolia testnets
- Deploy core contracts: NexusDelegation, AgentRegistry, NexusUSD, CrosschainBridge
- Deploy mock yield sources: MockAavePool, MockComet
- Update environment variables with deployed addresses
- Add DEPLOYMENT.md with all contract addresses and explorer links
- Fix type conversion bug in DeployYieldMocks.s.sol

Base Sepolia deployments:
- NexusDelegation: 0x74abaEC9aB23e08069751b3b9CA47Ce3FEa17d38
- AgentRegistry: 0x619AeC8bB48357D967F06F2d1582592455782B29
- NexusUSD: 0x1A54562813c35151a5e8cF4105221212ad97Bf52
- CrosschainBridge: 0x0f9ab104b1aB9dd5e5eA9b8280ad93147e282Cb0

OP Sepolia deployments:
- NexusDelegation: 0xAeb805427cb9BB4978c0EEf0FBe4bA64549cA85D
- AgentRegistry: 0x00DeFBE21758614CC5C2Ed6EB982E3084B952285
- NexusUSD: 0xaA6B7D1b89B05BEABD2F23baf0110806082D09a5
- CrosschainBridge: 0x3CfdE8a17d24bD93f64A38FC3FAd72C4F6585255

+++ contracts/script/DeployYieldMocks.s.sol
@@ -11,7 +11,7 @@ contract DeployYieldMocks is Script {
         address token = vm.envAddress("YIELD_TOKEN_ADDRESS");
         uint128 aaveLiquidityRate = uint128(vm.envOr("AAVE_LIQUIDITY_RATE", uint256(3e25)));
         uint256 cometUtilization = vm.envOr("COMET_UTILIZATION", uint256(8e17));
-        uint256 cometSupplyRate = vm.envOr("COMET_SUPPLY_RATE", uint256(1e9));
+        uint64 cometSupplyRate = uint64(vm.envOr("COMET_SUPPLY_RATE", uint256(1e9)));
 
         vm.startBroadcast(deployerPrivateKey);
